+++ package.json
@@ -25,7 +25,8 @@
     "react": "18.2.0",
     "react-cookie": "^4.1.1",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.41.0"
+    "react-hook-form": "^7.41.0",
+    "react-query": "^3.39.2"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.20.2",
