+++ src/libs/api/posts.ts
@@ -18,15 +18,6 @@ interface IGetPostByIdParams {
   userId?: number
 }
 
-export interface BookmarkType {
-  postId: number
-  title: string
-  nickname: string
-  createdAt: string
-  thumbnail: null
-  description: string
-} 
-
 export interface IPostType {
   id: number
   userId: number
